修复：将菜单插件添加到主界面

+++ GUI/script/compile_ui.py
@@ -2,6 +2,7 @@ from pathlib import Path
 import subprocess
 import shutil
 import re
+from typing import List
 
 CURRENT_DIR = Path(__file__).parent
 
@@ -10,35 +11,53 @@ def to_pascal_case(name: str) -> str:
     parts = re.findall(r"[A-Za-z0-9]+", name)
     return ''.join(p.capitalize() for p in parts if p)
 
-def build_all_ui():
-    # 清除所有ui_widget下的带有plugin结尾的文件夹
+def reset_ui_files():
+    """
+    重置已编译的ui文件夹
+    """
     ui_widget_path = CURRENT_DIR.parent / "plugins" / "ui_widget"
     if ui_widget_path.exists():
-        for f in ui_widget_path.rglob("*plugin"):
-            if f.is_dir():
-                print(f"删除非空文件夹（递归）：{f}")   
+        for f in ui_widget_path.iterdir():
+            if f.is_dir() and f.name.endswith("_plugin"):
+                print(f"删除ui文件夹：{f}")   
                 shutil.rmtree(f)
-    else:
-        print(f"ui_widget路径不存在：{ui_widget_path}")
+
+def create_ui_plugin():
+    """
+    创建ui插件文件夹
+    """
+    ui_widget_path = CURRENT_DIR.parent / "plugins" / "ui_widget"
     ui_path = CURRENT_DIR.parent / "resource" / "ui"
-    print(f"ui资源文件夹：{ui_path}\n")
+    for f in ui_path.iterdir():
+        if f.is_dir() and f.name.endswith("_plugin"):
+            (ui_widget_path / f"{f.name}").mkdir()
+            (ui_widget_path / f"{f.name}" / "__init__.py").touch() # 创建空的__init__.py文件
+
+def main():
+    """
+    编译UI文件
+    """
+    print("重置已编译的ui文件夹")
+    reset_ui_files()
+    create_ui_plugin()
+    print("编译UI文件")
+    ui_path = CURRENT_DIR.parent / "resource" / "ui"
+    ui_widget_path = CURRENT_DIR.parent / "plugins" / "ui_widget"
     for f in ui_path.rglob("*.ui"):
-        print(f"编译UI文件：{f}")
-        # 编译UI文件
-        out_parent = CURRENT_DIR.parent / 'plugins' / 'ui_widget' / f.parent.name
-        if not out_parent.exists():
-            out_parent.mkdir(parents=True)
-            open(out_parent / "__init__.py", "w").close() # 创建空的__init__.py文件
-        cmd = f"pyside6-uic {f} -o {out_parent / (f.stem + '.py')}"
-        f_init = open(out_parent / "__init__.py", "a") # 以追加模式打开__init__.py文件
-        # 将文件名转换为真正的大驼峰（PascalCase），如 main_window -> MainWindow
-        class_name = to_pascal_case(f.stem)
-        f_init.write(f"from .{f.stem} import Ui_{class_name}\n")
-        f_init.close()
-        print(f"执行命令：{cmd}")
-        subprocess.run(cmd, shell=True)
+        if f.is_file() and f.name.endswith(".ui"):
+            print(f"编译UI文件：{f}")
+            ui_build_path = ui_widget_path / f"{f.parent.name}" / f"ui_{f.stem}.py"
+            ui_build_path.parent.mkdir(parents=True, exist_ok=True)
+            subprocess.run([
+                "pyside6-uic",
+                str(f),
+                "-o",
+                str(ui_build_path),
+            ], check=True)
+            # 处理init文件
+            init_path = ui_build_path.parent / "__init__.py"
+            open(init_path, "a").write(f"from .{ui_build_path.name.split('.')[0]} import Ui_{to_pascal_case(f.stem)}\n")
+    print("编译完成")
 
 if __name__ == "__main__":
-    # 编译所有UI文件
-    print("编译所有UI文件...")
-    build_all_ui()
+    main()
